Create Player 2
In this commit, i fixed some earlier bugs which prevented the level from changing when the player was on the sugar. I also fixed the problem of when the character touches the trap(the fire)it won't reset the level. Now it does.

I also attempted to add another player, and made new levels however, i am trying to make the levels change when the characters are on the goals. Right now it is not working and in the future will be where I will start