Update last new file
i made a new level to explain the fragile tiles but ran into an issue. in that level, when the player runs into the fire, the previous tile they were on turns into a broken tile while the others reset to normal. I tried to fix this with some extra code in the trap section but that doesn't seem to be working